task(SDK-5141, SDK-5181) - Adds support for sticky and dismissAfter (#885)
* task(SDK-5141, SDK-5181) - Adds support for sticky and dismissAfter

* task(SDK-5141, SDK-5181) - Adds more tests

* task(SDK-5141, SDK-5181) - Coderabbit comments

* task(SDK-5141, SDK-5181) - Coderabbit comments

+++ clevertap-pushtemplates/src/main/java/com/clevertap/android/pushtemplates/content/TimerBigContentView.kt
@@ -7,7 +7,7 @@ import com.clevertap.android.pushtemplates.TimerTemplateData
 
 internal class TimerBigContentView(
     context: Context,
-    timer_end: Int?,
+    timer_end: Long?,
     renderer: TemplateRenderer,
     data: TimerTemplateData
 ) :
